+++ packages/hashing/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 1.14.0
 
-- First version of package beginning from dvote-js library version 1.13.2
+- First version of package, starting from dvote-js version 1.13.2